images/backend/requirements.txt
django
djangorestframework
channels
psycopg2-binary
dotenv
redis
django-cors-headers
django-extensions
django-rest-framework
djangorestframework-simplejwt
djangorestframework-jwt
django-rest-passwordreset
django-allauth
django-allauth-ui
requests
drf-yasg
Pillow
django-filter
coverage
pytest
pytest-django
python-dotenv
django-otp 
pyotp
